refactor(hub-router): split 1329-line main.go; dedupe packet helpers; vet/fmt clean
Pure code motion into bootstrap/http_handlers/tcp_proxy/udp_proxy/
dynamic_ports (all <400 lines); the JWT and target packet-parse helpers
existed in 4 copies — now one implementation in helpers.go with tests
(9 passing). Fixes pre-existing go vet IPv6 findings (JoinHostPort) and
gofmt drift across the service. TLS MinVersion 1.2 verified on the sole
tls.Config site.
