+++ services/hub-router/proxy/auth/provider.go
@@ -12,21 +12,21 @@
 package auth
 
 import (
-    "github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin"
 )
 
 type User struct {
-    ID       string                 `json:"id"`
-    Email    string                 `json:"email"`
-    Name     string                 `json:"name"`
-    Groups   []string               `json:"groups"`
-    Metadata map[string]interface{} `json:"metadata"`
+	ID       string                 `json:"id"`
+	Email    string                 `json:"email"`
+	Name     string                 `json:"name"`
+	Groups   []string               `json:"groups"`
+	Metadata map[string]interface{} `json:"metadata"`
 }
 
 type Provider interface {
-    LoginHandler() gin.HandlerFunc
-    CallbackHandler() gin.HandlerFunc
-    LogoutHandler() gin.HandlerFunc
-    ValidateToken(token string) (*User, error)
-    GetUser(ctx *gin.Context) (*User, error)
+	LoginHandler() gin.HandlerFunc
+	CallbackHandler() gin.HandlerFunc
+	LogoutHandler() gin.HandlerFunc
+	ValidateToken(token string) (*User, error)
+	GetUser(ctx *gin.Context) (*User, error)
 }
